requirements.txt
Django==5.1
django_markdownx==4.0.7
django_mdeditor==0.1.20
django_taggit==6.0.0
djangorestframework==3.15.2
django-cors-headers
psycopg2-binary